Refactor ReportGenerator.java to storage package
