+++ src/main/java/seedu/address/logic/commands/GenReportsCommand.java
@@ -7,7 +7,7 @@ import java.util.List;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.entity.body.Body;
-import seedu.address.report.ReportGenerator;
+import seedu.address.storage.ReportGenerator;
 
 //@@author bernicechio
 /**
